requirements.txt
streamlit
openai
faiss-cpu
langchain
requests
configparser
